{"schema":"decodex/commit/1","summary":"Split docs OKF and autonomy proposal modules","authority":"manual"}
