+++ apps/decodex/src/autonomy_proposal/validation.rs
@@ -1,5 +1,4 @@
-#[allow(clippy::wildcard_imports)]
-use super::*;
+#[allow(clippy::wildcard_imports)] use super::*;
 
 pub(super) fn proposal_refusals(
 	objective: Option<&AutonomyObjectiveContract>,
